+++ RELEASE_NOTES.md
@@ -3,6 +3,8 @@ Release notes
 
 ### Unreleased
 
+* Added `HttpResponseMessage` assertion `HaveStringContentSatisfying`. Note that this is async and therefore not
+  chainable.
 * The `HttpResponseMessage` assertions `HaveHeader` and `HaveHeaderValue` now correctly detects content headers such
   as `Content-Type` (which are set on `HttpContent` and not `HttpResponseMessage`).